BOT_Detector/requirements.txt
flask==3.0.3
flask-cors==4.0.0
langchain==0.2.16
langchain-google-genai==1.0.10
google-generativeai==0.7.2
python-dotenv==1.0.1
